// interfaces/reports.interface.ts
export interface ReportOtpRequestInterface {
    _id?: any;
    createdAt?: string;
    mobile?: string;
    otp?: string;
    firmName?: string;
    contactPerson?: string;
    state?: string;
    city?: string;
    address?: string;
    postalCode?: string;
    appVersion?: string;
    deviceType? : string;
    deviceName? : string;
    loginAt? : string
}